pyproject.toml
[project]
name = "veadk-examples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "veadk-python>=0.2.5",
    "opensearch-py>=2.8.0",
    "tos>=2.8.5",
    "fastmcp>=2.12.0",
    "mcp-server-time>=2025.8.4",
    "litellm<=1.75", # 高版本有个loggingworker的bug，先降一下
    "volcengine>=1.0.198",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]

[tool.uv.sources]
veadk-python = { git = "https://github.com/volcengine/veadk-python.git" }
